requirements.txt
mysql-connector-python==9.1.0
PyMySQL==1.1.1
tabulate==0.9.0
